requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-property==0.2.2
async-timeout==4.0.3
asyncstdlib==3.12.0
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
browser-cookie3==0.19.1
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.2.1
curl-cffi==0.5.10
duckduckgo_search==4.1.1
exceptiongroup==1.2.0
fastapi==0.108.0
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.1
fsspec==2023.12.2
g4f==0.1.9.8
h11==0.14.0
html2text==2020.1.16
huggingface-hub==0.20.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
Js2Py==0.74
keyboard==0.13.5
loguru==0.7.2
lxml==5.0.1
lz4==4.3.3
MarkupSafe==2.1.3
mmh3==4.0.1
mtranslate==1.8
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.3
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
platformdirs==4.1.0
psutil==5.9.7
py-arkose-generator==0.0.0.2
pycparser==2.21
pycryptodome==3.19.1
pycryptodomex==3.19.1
pydantic==2.5.3
pydantic_core==2.14.6
PyExecJS==1.5.1
pygame==2.5.2
PyGetWindow==0.0.9
pyjsparser==2.7.1
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PySocks==1.7.1
python-dotenv==1.0.0
python-pptx==0.6.23
pyttsx3==2.90
pywin32==306
PyYAML==6.0.1
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.5
starlette==0.32.0.post1
tqdm==4.66.1
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
undetected-chromedriver==3.5.4
urllib3==2.1.0
uvicorn==0.25.0
waitress==2.1.2
webdriver-manager==4.0.1
websockets==12.0
Werkzeug==3.0.1
win32-setctime==1.1.0
wsproto==1.2.0
XlsxWriter==3.1.9
yarl==1.9.4
youtube-transcript-api==0.6.2
